Folder structure for contribution
